Restructuring test packages in line with main package structure
