+++ location4j/src/test/java/com/tomaytotomato/location4j/usecase/lookup/FindCityTest.java
@@ -1,11 +1,9 @@
-package com.tomaytotomato.location4j.usecase;
+package com.tomaytotomato.location4j.usecase.lookup;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.tomaytotomato.location4j.usecase.lookup.LocationService;
-import com.tomaytotomato.location4j.usecase.lookup.FindCity;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
